build: 重构构建配置并迁移至 tsdown
- 移除 tsconfig.module.json 文件
- 新增 tsdown.config.ts 文件，配置 tsdown 构建工具
- 删除 tsup.config.ts 和 tsup.modules.ts 文件，移除 tsup 构建工具
- 更新构建选项，包括入口文件、格式、依赖处理等
- 添加复制文件和清理目录的钩子函数
