+++ .vscode/settings.json
@@ -7,5 +7,12 @@
   "editor.formatOnSaveMode": "file",
   "editor.codeActionsOnSave": {
     "source.fixAll.eslint": "always"
+  },
+  "files.autoSave": "afterDelay",
+  "explorer.fileNesting.enabled": true,
+  "explorer.fileNesting.patterns": {
+    ".env": "*.env",
+    "tsconfig.json": "tsconfig.*.json, env.d.ts, tsup.*.ts, tsdown.*.ts",
+    "package.json": "package-lock.json, pnpm*, .yarnrc*, yarn*, .eslint*, eslint*, .prettier*, prettier*, .editorconfig, CHANGELOG*.md, CODE_OF_CONDUCT.md, *.gitignore, LICENSE, README.md, .release*.json, .gitattributes"
   }
 }
