--- tsconfig.module.json
@@ -1,7 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./dist/exports",
-    "rootDir": "./src/exports",
-  }
-}
